% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{get_var_simulations}
\alias{get_var_simulations}
\title{get_discrete_simulations}
\usage{
get_var_simulations(
  num_years = 5,
  num_paths = 10,
  frequency = "quarter",
  perc_change = FALSE,
  return_sdf = FALSE
)
}
\arguments{
\item{num_years}{Number of years to forecast, counting from 2021-01-01.
Default is 5 years, recommended period is less than 10 years.}

\item{num_paths}{Number of simulation paths. Default is 10 paths.}

\item{frequency}{One of "year", "quarter", and "month". Default is "quarter",
which is the simulation frequency for the Vector Autoregression model. Linear
interpolation will be used if the required frequency is higher, whereas
arithmetic average will be used if the frequency is lower.}

\item{perc_change}{If the outputs are expressed in terms of percentage
change.Default is FALSE}

\item{return_sdf}{If the VAR-based stochastic discount factors are returned.
Default is FALSE.}
}
\value{
A list containing 10 data frames for the simulated trajectories for
each economic variable, and a list of white noises in the VAR model.
}
\description{
Returns the simulated paths for various economic and financial variables:
(1) Australia 3-month zero-coupon yields, (2) Australia 10-year zero-coupon
spread, (3) New South Wales houses value index, (4) New South Wales houses
rental yields, (5) Australian GDP, (6) Australian CPI, (7) S&P/ASX200 closing
price, (8) Australian dollar trade-weighted index, (9) Australia mortgage
rate, (10) New South Wales unemployment rate (in \%).
Simulations are based on a Vector Autoregression model. This function uses
the package \code{zoo} to convert the frequnency units.
}
\examples{
sim = get_discrete_simulations(num_years = 10, num_paths = 100,
frequency = "year", return_sdf = T). To obtain all trajectories of Australia
3-month zero-coupon yields, type sim$zcp3m_yield, to obtain the noises in the
first trajectory, type sim$noise$trajectory_1.

}
